Fix source duration precision bug

+++ src/kitsource.c
@@ -270,7 +270,7 @@ int Kit_GetSourceStreamCount(const Kit_Source *src) {
 
 double Kit_GetSourceDuration(const Kit_Source *src) {
     assert(src != NULL);
-    return (((AVFormatContext *)src->format_ctx)->duration / AV_TIME_BASE);
+    return ((AVFormatContext *)src->format_ctx)->duration / (double)AV_TIME_BASE;
 }
 
 int Kit_GetSourceStreamList(const Kit_Source *src, const Kit_StreamType type, int *list, int size) {
